refactor(core): Escape html comment in i18n_parse
Should someone need to embed a compiled angular app inline in a script
tag in an HTML page, the unescaped HTML comment will at best change the
behavior of this code, and at worst be rejected by the sanitizer of the
template engine used to generate the page. Escaping it should prevent
this, unless the minifier is so aggressive that it removes the escapes.